moved to fuels 0.46

+++ contracts/market/tests/utils/contracts_utils/token_utils.rs
@@ -4,14 +4,14 @@ use fuels::{
 };
 use src20_sdk::{DeployTokenConfig, TokenContract};
 
-
 pub struct Asset {
     pub config: DeployTokenConfig,
     pub contract_id: ContractId,
     pub asset_id: AssetId,
     pub instance: Option<TokenContract<WalletUnlocked>>,
     pub default_price: u64,
-    pub decimals: u8,
+    pub decimals: u64,
     pub symbol: String,
     pub coingeco_id: String,
+    // pub asset_config: AssetConfig,
 }
